+++ bookkeeping-app/__tests__/components/contexts/PropertiesCtx.test.jsx
@@ -2,9 +2,9 @@ import { render, screen, fireEvent, waitFor } from "@testing-library/react";
 import { useContext } from "react";
 import useSWRImmutable from "swr/immutable";
 
-import { PropertiesCtxProvider } from "@/src/components/contexts/PropertiesCtx";
-import PropertiesCtx from "@/src/components/contexts/PropertiesCtx";
-import AuthCtx from "@/src/components/contexts/AuthCtx";
+import { PropertiesCtxProvider } from "@/src/contexts/PropertiesCtx";
+import PropertiesCtx from "@/src/contexts/PropertiesCtx";
+import AuthCtx from "@/src/contexts/AuthCtx";
 import { api, ApiError } from "@/src/Client";
 
 jest.mock("swr/immutable", () => ({
@@ -33,7 +33,7 @@ jest.mock("@/src/Client", () => {
 });
 
 const mockShowToast = jest.fn();
-jest.mock("@/src/components/contexts/ToastCtx", () => ({
+jest.mock("@/src/contexts/ToastCtx", () => ({
     useToast: () => ({ showToast: mockShowToast }),
 }));
 
@@ -111,3 +111,4 @@ describe("PropertiesCtx", () => {
         await waitFor(() => expect(mockShowToast).toHaveBeenCalledWith("Error updating Property", "error", 5000));
     });
 });
+
