+++ bookkeeping-app/src/components/elements/dropdowns/EntityDropdown.jsx
@@ -1,7 +1,7 @@
 import classes from "./EntityDropdown.module.css";
 
 import { useState, useContext, useEffect, useRef } from "react";
-import EntitiesCtx from "../../contexts/EntitiesCtx";
+import EntitiesCtx from "../../../contexts/EntitiesCtx";
 
 import upChevIconB from "../../../assets/chevron-up-icon.svg";
 import downChevIconB from "../../../assets/chevron-down-icon.svg";
