+++ bookkeeping-app/src/contexts/TransactionsCtx.jsx
@@ -1,7 +1,7 @@
 import { createContext, useState, useContext } from "react";
 import useSWRImmutable from "swr/immutable";
 
-import { ApiError, api } from "../../Client";
+import { ApiError, api } from "../Client";
 import { useToast } from "./ToastCtx";
 import AccountsCtx from "./AccountsCtx";
 import EntitiesCtx from "./EntitiesCtx";
